packages/destatis/pyproject.toml
[project]
name = "parsimony-destatis"
version = "0.4.0"
description = "Destatis (German federal statistics) connector for the parsimony framework"
authors = [{ name = "Ockham.sh", email = "team@ockham.sh" }]
license = "Apache-2.0"
readme = "README.md"
requires-python = ">=3.11"
keywords = ["finance", "data", "connectors", "parsimony", "destatis"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Financial and Insurance Industry",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Office/Business :: Financial",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Typing :: Typed",
]
dependencies = [
    "parsimony-core>=0.4.0,<0.5",
    "pydantic>=2.11.1,<3",
    "pandas>=2.3.0,<3",
]

[project.optional-dependencies]
# Heavy deps for running the HF catalog publisher under scripts/. Kept
# behind an extra so `pip install parsimony-destatis` stays lean — only
# catalog maintainers ever need parsimony-core[standard-onnx] (faiss +
# onnxruntime + sentence-transformers).
publish = [
    "parsimony-core[standard-onnx]>=0.4.0,<0.5",
]
dev = [
    "pytest>=9.0.3",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=5.0",
    "respx>=0.22.0",
    "ruff>=0.15.10",
    "mypy>=1.10",
]

[project.urls]
Homepage = "https://www.destatis.de"
Repository = "https://github.com/ockham-sh/parsimony-connectors"
Issues = "https://github.com/ockham-sh/parsimony-connectors/issues"

[project.entry-points."parsimony.providers"]
destatis = "parsimony_destatis"


[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["parsimony_destatis"]

[tool.hatch.build.targets.sdist]
include = ["parsimony_destatis", "README.md", "LICENSE", "CHANGELOG.md"]

[tool.ruff]
target-version = "py311"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]

[tool.ruff.lint.per-file-ignores]
# Publish scripts call logging.basicConfig() before importing parsimony so
# the kernel's INFO logs surface during the long-running publish pipeline.
"scripts/*" = ["E402"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_ignores = true
ignore_missing_imports = true

[tool.pytest.ini_options]
addopts = "--import-mode=importlib -m 'not integration'"
asyncio_mode = "auto"
markers = [
    "integration: hits live APIs (may be slow, requires env vars)",
]
